Provider cleanup work
