+++ internal/mysql/provider/rds/provider.go
@@ -6,14 +6,18 @@ import (
 	"log"
 	"strings"
 
-	"github.com/skpr/mtk/internal/mysql/providers"
+	"github.com/skpr/mtk/internal/mysql/provider"
+	providerutils "github.com/skpr/mtk/internal/mysql/provider/utils"
 )
 
 // Client used for dumping a database and/or table.
 type Client struct {
-	providers.MTKProvider
+	provider.Interface
 	DB     *sql.DB
 	Logger *log.Logger
+
+	DataPath string // S3 URI configuration
+	Region   string // Region configuration
 }
 
 // NewClient for dumping a full or single table from a database.
@@ -24,37 +28,9 @@ func NewClient(db *sql.DB, logger *log.Logger) *Client {
 	}
 }
 
-// QueryColumnsForTable for a given table.
-func (d *Client) QueryColumnsForTable(table string, params providers.DumpParams) ([]string, error) {
-	var rows *sql.Rows
-
-	rows, err := d.DB.Query(fmt.Sprintf("SELECT * FROM `%s` LIMIT 1", table))
-	if err != nil {
-		return nil, err
-	}
-
-	defer rows.Close()
-
-	columns, err := rows.Columns()
-	if err != nil {
-		return nil, err
-	}
-
-	for k, column := range columns {
-		replacement, ok := params.SelectMap[strings.ToLower(table)][strings.ToLower(column)]
-		if ok {
-			columns[k] = fmt.Sprintf("%s AS `%s`", replacement, column)
-		} else {
-			columns[k] = fmt.Sprintf("`%s`", column)
-		}
-	}
-
-	return columns, nil
-}
-
 // GetSelectQueryForTable will return a complete SELECT query to export data from a table.
-func (d *Client) GetSelectQueryForTable(table string, params providers.DumpParams) (string, error) {
-	cols, err := d.QueryColumnsForTable(table, params)
+func (d *Client) GetSelectQueryForTable(table string, params provider.DumpParams) (string, error) {
+	cols, err := providerutils.QueryColumnsForTable(d.DB, table, params)
 	if err != nil {
 		return "", err
 	}
@@ -66,12 +42,12 @@ func (d *Client) GetSelectQueryForTable(table string, params providers.DumpParam
 		query = fmt.Sprintf("%s WHERE %s", query, where)
 	}
 
-	query = fmt.Sprintf("%s INTO OUTFILE S3 '%s/%s.csv'", query, params.DataPath, table)
+	query = fmt.Sprintf("%s INTO OUTFILE S3 '%s/%s.csv'", query, d.DataPath, table)
 	query = fmt.Sprintf("%s FIELDS TERMINATED BY ',' ENCLOSED BY '\"' LINES TERMINATED BY '\\n'", query)
 	query = fmt.Sprintf("%s MANIFEST ON", query)
 	query = fmt.Sprintf("%s OVERWRITE ON", query)
 
-	importQuery, err := d.GetLoadQueryForTable(table, params.DataPath, params.Region)
+	importQuery, err := d.GetLoadQueryForTable(table, d.DataPath, d.Region)
 	if err != nil {
 		return "", err
 	}
