+++ internal/mysql/tables_test.go
@@ -11,12 +11,12 @@ import (
 	"github.com/stretchr/testify/assert"
 
 	"github.com/skpr/mtk/internal/mysql/mock"
-	"github.com/skpr/mtk/internal/mysql/providers"
+	"github.com/skpr/mtk/internal/mysql/provider"
 )
 
 func TestMySQLFlushTable(t *testing.T) {
 	db, mock := mock.GetDB(t)
-	dumper := NewClient(db, log.New(os.Stdout, "", 0))
+	dumper := NewClient(db, log.New(os.Stdout, "", 0), "")
 	mock.ExpectExec("FLUSH TABLES `table`").WillReturnResult(sqlmock.NewResult(0, 1))
 	_, err := dumper.FlushTable("table")
 	assert.Nil(t, err)
@@ -24,7 +24,7 @@ func TestMySQLFlushTable(t *testing.T) {
 
 func TestMySQLUnlockTables(t *testing.T) {
 	db, mock := mock.GetDB(t)
-	dumper := NewClient(db, log.New(os.Stdout, "", 0))
+	dumper := NewClient(db, log.New(os.Stdout, "", 0), "")
 	mock.ExpectExec("UNLOCK TABLES").WillReturnResult(sqlmock.NewResult(0, 1))
 	_, err := dumper.UnlockTables()
 	assert.Nil(t, err)
@@ -32,7 +32,7 @@ func TestMySQLUnlockTables(t *testing.T) {
 
 func TestMySQLQueryTables(t *testing.T) {
 	db, mock := mock.GetDB(t)
-	dumper := NewClient(db, log.New(os.Stdout, "", 0))
+	dumper := NewClient(db, log.New(os.Stdout, "", 0), "")
 	mock.ExpectQuery("SHOW FULL TABLES").WillReturnRows(
 		sqlmock.NewRows([]string{"Tables_in_database", "Table_type"}).
 			AddRow("table1", "BASE TABLE").
@@ -45,7 +45,7 @@ func TestMySQLQueryTables(t *testing.T) {
 
 func TestMySQLLockTableRead(t *testing.T) {
 	db, mock := mock.GetDB(t)
-	dumper := NewClient(db, log.New(os.Stdout, "", 0))
+	dumper := NewClient(db, log.New(os.Stdout, "", 0), "")
 	mock.ExpectExec("LOCK TABLES `table` READ").WillReturnResult(sqlmock.NewResult(0, 1))
 	_, err := dumper.LockTableReading("table")
 	assert.Nil(t, err)
@@ -53,7 +53,7 @@ func TestMySQLLockTableRead(t *testing.T) {
 
 func TestMySQLGetTablesHandlingErrorWhenListingTables(t *testing.T) {
 	db, mock := mock.GetDB(t)
-	dumper := NewClient(db, log.New(os.Stdout, "", 0))
+	dumper := NewClient(db, log.New(os.Stdout, "", 0), "")
 	expectedErr := errors.New("broken")
 	mock.ExpectQuery("SHOW FULL TABLES").WillReturnError(expectedErr)
 	tables, err := dumper.QueryTables()
@@ -63,7 +63,7 @@ func TestMySQLGetTablesHandlingErrorWhenListingTables(t *testing.T) {
 
 func TestMySQLGetTablesHandlingErrorWhenScanningRow(t *testing.T) {
 	db, mock := mock.GetDB(t)
-	dumper := NewClient(db, log.New(os.Stdout, "", 0))
+	dumper := NewClient(db, log.New(os.Stdout, "", 0), "")
 	mock.ExpectQuery("SHOW FULL TABLES").WillReturnRows(
 		sqlmock.NewRows([]string{"Tables_in_database", "Table_type"}).AddRow(1, nil))
 	tables, err := dumper.QueryTables()
@@ -78,7 +78,7 @@ func TestMySQLDumpCreateTable(t *testing.T) {
 		"PRIMARY KEY (`id`), KEY `idx_name` (`name`) " +
 		") ENGINE=InnoDB AUTO_INCREMENT=1 DEFAULT CHARSET=utf8"
 	db, mock := mock.GetDB(t)
-	dumper := NewClient(db, log.New(os.Stdout, "", 0))
+	dumper := NewClient(db, log.New(os.Stdout, "", 0), "")
 	mock.ExpectQuery("SHOW CREATE TABLE `table`").WillReturnRows(
 		sqlmock.NewRows([]string{"Table", "Create Table"}).
 			AddRow("table", ddl),
@@ -91,7 +91,7 @@ func TestMySQLDumpCreateTable(t *testing.T) {
 
 func TestMySQLDumpCreateTableHandlingErrorWhenScanningRows(t *testing.T) {
 	db, mock := mock.GetDB(t)
-	dumper := NewClient(db, log.New(os.Stdout, "", 0))
+	dumper := NewClient(db, log.New(os.Stdout, "", 0), "")
 	mock.ExpectQuery("SHOW CREATE TABLE `table`").WillReturnRows(
 		sqlmock.NewRows([]string{"Table", "Create Table"}).AddRow("table", nil))
 	buffer := bytes.NewBuffer(make([]byte, 0))
@@ -100,10 +100,10 @@ func TestMySQLDumpCreateTableHandlingErrorWhenScanningRows(t *testing.T) {
 
 func TestMySQLGetRowCount(t *testing.T) {
 	db, mock := mock.GetDB(t)
-	dumper := NewClient(db, log.New(os.Stdout, "", 0))
+	dumper := NewClient(db, log.New(os.Stdout, "", 0), "")
 	mock.ExpectQuery("SELECT COUNT\\(\\*\\) FROM `table` WHERE c1 > 0").WillReturnRows(
 		sqlmock.NewRows([]string{"COUNT(*)"}).AddRow(1234))
-	count, err := dumper.GetRowCountForTable("table", providers.DumpParams{
+	count, err := dumper.GetRowCountForTable("table", provider.DumpParams{
 		WhereMap: map[string]string{"table": "c1 > 0"},
 	})
 	assert.Nil(t, err)
@@ -112,10 +112,10 @@ func TestMySQLGetRowCount(t *testing.T) {
 
 func TestMySQLGetRowCountHandlingError(t *testing.T) {
 	db, mock := mock.GetDB(t)
-	dumper := NewClient(db, log.New(os.Stdout, "", 0))
+	dumper := NewClient(db, log.New(os.Stdout, "", 0), "")
 	mock.ExpectQuery("SELECT COUNT\\(\\*\\) FROM `table` WHERE c1 > 0").WillReturnRows(
 		sqlmock.NewRows([]string{"COUNT(*)"}).AddRow(nil))
-	count, err := dumper.GetRowCountForTable("table", providers.DumpParams{
+	count, err := dumper.GetRowCountForTable("table", provider.DumpParams{
 		WhereMap: map[string]string{"table": "c1 > 0"},
 	})
 	assert.NotNil(t, err)
