+++ internal/mysql/write_test.go
@@ -12,16 +12,16 @@ import (
 	"github.com/stretchr/testify/assert"
 
 	"github.com/skpr/mtk/internal/mysql/mock"
-	"github.com/skpr/mtk/internal/mysql/providers"
+	"github.com/skpr/mtk/internal/mysql/provider"
 )
 
 func TestMySQLDumpTableHeader(t *testing.T) {
 	db, mock := mock.GetDB(t)
-	dumper := NewClient(db, log.New(os.Stdout, "", 0))
+	dumper := NewClient(db, log.New(os.Stdout, "", 0), "stdout")
 	mock.ExpectQuery("SELECT COUNT\\(\\*\\) FROM `table`").WillReturnRows(
 		sqlmock.NewRows([]string{"COUNT(*)"}).AddRow(1234))
 	buffer := bytes.NewBuffer(make([]byte, 0))
-	count, err := dumper.WriteTableHeader(buffer, "table", providers.DumpParams{})
+	count, err := dumper.WriteTableHeader(buffer, "table", provider.DumpParams{})
 	assert.Equal(t, uint64(1234), count)
 	assert.Nil(t, err)
 	assert.Contains(t, buffer.String(), "Data for table `table`")
@@ -30,25 +30,25 @@ func TestMySQLDumpTableHeader(t *testing.T) {
 
 func TestMySQLDumpTableHeaderHandlingError(t *testing.T) {
 	db, mock := mock.GetDB(t)
-	dumper := NewClient(db, log.New(os.Stdout, "", 0))
+	dumper := NewClient(db, log.New(os.Stdout, "", 0), "stdout")
 	mock.ExpectQuery("SELECT COUNT\\(\\*\\) FROM `table`").WillReturnRows(
 		sqlmock.NewRows([]string{"COUNT(*)"}).AddRow(nil))
 	buffer := bytes.NewBuffer(make([]byte, 0))
-	count, err := dumper.WriteTableHeader(buffer, "table", providers.DumpParams{})
+	count, err := dumper.WriteTableHeader(buffer, "table", provider.DumpParams{})
 	assert.Equal(t, uint64(0), count)
 	assert.NotNil(t, err)
 }
 
 func TestMySQLDumpTableLockWrite(t *testing.T) {
 	buffer := bytes.NewBuffer(make([]byte, 0))
-	dumper := NewClient(nil, log.New(os.Stdout, "", 0))
+	dumper := NewClient(nil, log.New(os.Stdout, "", 0), "stdout")
 	dumper.WriteTableLockWrite(buffer, "table")
 	assert.Contains(t, buffer.String(), "LOCK TABLES `table` WRITE;")
 }
 
 func TestMySQLDumpUnlockTables(t *testing.T) {
 	buffer := bytes.NewBuffer(make([]byte, 0))
-	dumper := NewClient(nil, log.New(os.Stdout, "", 0))
+	dumper := NewClient(nil, log.New(os.Stdout, "", 0), "stdout")
 	dumper.WriteUnlockTables(buffer)
 	assert.Contains(t, buffer.String(), "UNLOCK TABLES;")
 }
@@ -56,7 +56,7 @@ func TestMySQLDumpUnlockTables(t *testing.T) {
 func TestMySQLDumpTableData(t *testing.T) {
 	db, mock := mock.GetDB(t)
 	buffer := bytes.NewBuffer(make([]byte, 0))
-	dumper := NewClient(db, log.New(os.Stdout, "", 0))
+	dumper := NewClient(db, log.New(os.Stdout, "", 0), "stdout")
 
 	mock.ExpectQuery("SELECT \\* FROM `table` LIMIT 1").WillReturnRows(
 		sqlmock.NewRows([]string{"id", "language"}).
@@ -71,10 +71,8 @@ func TestMySQLDumpTableData(t *testing.T) {
 			AddRow(5, "Rust").
 			AddRow(6, "Closure"))
 
-	assert.Nil(t, dumper.WriteTableData(buffer, "table", providers.DumpParams{
-		ExtendedInsertRows: 2,
-		Provider:           "stdout",
-	}))
+	assert.Nil(t, dumper.WriteTableData(buffer, "table", provider.DumpParams{
+		ExtendedInsertRows: 2}))
 
 	assert.Equal(t, strings.Count(buffer.String(), "INSERT INTO `table` VALUES"), 3)
 	assert.Equal(t, buffer.String(), "INSERT INTO `table` VALUES (1,'Go'),(2,'Java');\nINSERT INTO `table` VALUES (3,'C'),(4,'C++');\nINSERT INTO `table` VALUES (5,'Rust'),(6,'Closure');\n")
@@ -83,10 +81,8 @@ func TestMySQLDumpTableData(t *testing.T) {
 func TestMySQLDumpTableDataHandlingErrorFromSelectAllDataFor(t *testing.T) {
 	db, mock := mock.GetDB(t)
 	buffer := bytes.NewBuffer(make([]byte, 0))
-	dumper := NewClient(db, log.New(os.Stdout, "", 0))
+	dumper := NewClient(db, log.New(os.Stdout, "", 0), "stdout")
 	error := errors.New("fail")
 	mock.ExpectQuery("SELECT \\* FROM `table` LIMIT 1").WillReturnError(error)
-	assert.Equal(t, error, dumper.WriteTableData(buffer, "table", providers.DumpParams{
-		Provider: "stdout",
-	}))
+	assert.Equal(t, error, dumper.WriteTableData(buffer, "table", provider.DumpParams{}))
 }
